+++ clang-tools-extra/clang-tidy/misc/YardenCommentSpacebeforeblockCheck.h
@@ -1,33 +1,21 @@
-//===--- YardenCommentSpacebeforeblockCheck.h - clang-tidy ------*- C++ -*-===//
-//
-// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-//===----------------------------------------------------------------------===//
-
 #ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MISC_YARDENCOMMENTSPACEBEFOREBLOCKCHECK_H
 #define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MISC_YARDENCOMMENTSPACEBEFOREBLOCKCHECK_H
 
 #include "../ClangTidyCheck.h"
 
-namespace clang::tidy::misc {
+namespace clang {
+namespace tidy {
+namespace misc {
 
-/// FIXME: Write a short description.
-///
-/// For the user-facing documentation see:
-/// http://clang.llvm.org/extra/clang-tidy/checks/misc/yarden-comment-spacebeforeblock.html
 class YardenCommentSpacebeforeblockCheck : public ClangTidyCheck {
 public:
-  YardenCommentSpacebeforeblockCheck(StringRef Name, ClangTidyContext *Context)
-      : ClangTidyCheck(Name, Context) {}
+  YardenCommentSpacebeforeblockCheck(StringRef Name, ClangTidyContext *Context);
   void registerMatchers(ast_matchers::MatchFinder *Finder) override;
   void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
-  bool isLanguageVersionSupported(const LangOptions &LangOpts) const override {
-    return LangOpts.CPlusPlus;
-  }
 };
 
-} // namespace clang::tidy::misc
+} // namespace misc
+} // namespace tidy
+} // namespace clang
 
 #endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_MISC_YARDENCOMMENTSPACEBEFOREBLOCKCHECK_H